# optimize addAll methods for ArrayTagSet